updatedel try

+++ src/domain/tests/tests.repo.js
@@ -5,26 +5,62 @@ class TestRepository {
     this.prisma = new PrismaClient();
   }
 
-  createTest(body) {
-    return this.prisma.test.create({
-      data: body,
-    });
+  async createTest(body) {
+    try {
+      return this.prisma.test.create({
+        data: body,
+      });
+    } catch (error) {
+      console.log(error);
+    }
   }
 
-  getTests() {
-    return this.prisma.test.findMany();
+  async getTests() {
+    try {
+      return this.prisma.test.findMany();
+    } catch (error) {
+      console.log(error);
+    }
   }
 
-  getTest(id) {
-    return this.prisma.test.findFirst({ where: { id } });
+  async getTest(id) {
+    try {
+      return this.prisma.test.findFirst({ where: { id } });
+    } catch (error) {
+      console.log(error);
+    }
+    
   }
 
-  deleteTest(id) {
-    return this.prisma.test.delete({ where: { id } });
+  async deleteTest(id) {
+    try {
+      const findTest = await this.prisma.test.findFirst({ where: { id } });
+      if (findTest) {
+        return this.prisma.test.delete({ where: { id } });
+      } else {
+        return { message: "User not found " };
+      }
+    } catch (error) {
+      console.log(error);
+    }
   }
 
-  updateTest(id, body) {
-    return this.prisma.test.update({ where: { id }, data: body });
+  async updateTest(id, data) {
+    try {
+      return this.prisma.test.update({
+        data: {
+          name: data.name,
+          shortName: data.shortName,
+          isActive: data.isActive,
+          price: data.price,
+          measurement: data.measurement,
+          updatedAt: data.updatedAt,
+        },
+        where: { id },
+      });
+    } catch (error) {
+      console.log(error);
+    }
   }
 }
